% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlot-package.R
\docType{package}
\name{xlot-package}
\alias{xlot}
\alias{xlot-package}
\title{xlot: Across-the-lot testing of an ELISA plate production run}
\description{
A Shiny app for exploring test data of an entire ELISA plate production run.
}
\author{
\strong{Maintainer}: Jeffrey Stevens \email{first.last@example.com}

}
\keyword{internal}
